tools/nut-usbinfo.pl: Reindent upower rules header
This piece of the code was really confusing because it looked like, at a
glance (and at more than a glance), that the line which we'd have read
in the generated rules file was commented out.

Turns out that beginning of lines in the generated rules could be in the
middle of the line in the code that generated it. Reindent everything so
that each line in the output corresponds to one line in the input,
except for empty lines.

There are no differences in output between the versions before and after
this change.